requirements.txt
yfinance==0.2.40
streamlit
pandas
plotly
anthropic
python-dotenv
kaleido
